refactor: rename return fields
