+++ src/types/index.ts
@@ -1,8 +1,7 @@
-import { PROTOCOL_ID } from "../visualizer";
-
 export * as blurTypes from "./blur";
 export * as looksrareTypes from "./looksrare";
 export * as seaportTypes from "./seaport";
+export * as visualizer from "./visualizer";
 
 export enum ASSET_TYPE {
   NATIVE = "NATIVE",
@@ -10,13 +9,6 @@ export enum ASSET_TYPE {
   ERC721 = "ERC721",
   ERC1155 = "ERC1155",
 }
-
-export type Domain = {
-  verifyingContract: string;
-  name: string;
-  chainId: string;
-  version: string;
-};
 
 export type Asset = {
   address: string;
@@ -27,37 +19,6 @@ export type Asset = {
 export type AssetInOut = {
   amounts: string[];
 } & Asset;
-
-type Liveness = {
-  from: number;
-  to: number;
-};
-
-type Approval = {
-  amounts: string[];
-  operator: string;
-  owner: string;
-  deadline?: number;
-} & Asset;
-
-export type VisualizationResult = {
-  protocol: PROTOCOL_ID;
-  liveness?: Liveness;
-  assetIn: AssetInOut[];
-  assetOut: AssetInOut[];
-  approval: Approval[];
-};
-
-export type EIP712Protocol<T> = {
-  /**
-   * @param {T} message EIP-712 message
-   * @param {Domain} domain EIP-712 domain
-   * @returns {VisualizationResult} assets impact and message liveness
-   * @throws {Error}
-   */
-  visualize: (message: T, domain: Domain) => VisualizationResult;
-  isCorrectDomain: (domain: Domain) => boolean;
-};
 
 export type PermitMessage = {
   owner?: string;
